pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "flask>=3.1.2",
    "fpdf>=1.7.2",
    "numpy>=2.4.0",
    "opencv-python>=4.11.0.86",
    "opencv-python-headless>=4.11.0.86",
    "pillow>=12.1.0",
    "tensorflow>=2.20.0",
    "werkzeug>=3.1.4",
]
